Home to Git 5/28/20

+++ APCS-A-2020/src/Tanks/Locatable.java
@@ -3,10 +3,10 @@ package Tanks;
 // Used from Pong project
 public interface Locatable 
 {
-    public void setPos( int x, int y);
-    public void setX( int x );
-    public void setY( int y );
+    public void setPos( double x, double y);
+    public void setX( double x );
+    public void setY( double y );
 
-    public int getX();
-    public int getY();
+    public double getX();
+    public double getY();
 }
